rename pagepartials -> pagetemplates
